[ANE-611] Unarchiving VSI scan (#6)
# Overview

This PR uses the internal `archive` library to recursively expand
archives while performing a VSI scan.

## Acceptance criteria

- Archives are expanded and fingerprinted while running
`vsi::scan::artifacts`.

## Testing plan

Added an automated integration test for this. The data was stolen from
the archive expansion tests.

## Risks

Some placeholder filter functionality was removed, as it now resides in
the archive library, and will be implemented as an extension of the
`archive::expand::walk` function. This should be double-checked, as I
was not sure this was entirely the correct move.

## References

ANE-611

## Checklist

- [ ] I added tests for this PR's change (or explained in the PR
description why tests don't make sense).