Fix up function part extraction (#18)
# Overview

I was getting panics on a code file in an actual project, the test code
added to this PR.

This was because the code to extract function signatures and bodies was
too abstracted and was getting tripped up on which parts of the function
belonged to which. This PR fixes that.

It also does some simplification on the `SnippetContext` type to reduce
possibility of overflow errors there. I'm not sure whether these were
also occurring, but I think it was possible.

## Acceptance criteria

More complicated function signatures now properly parse.

## Testing plan

I relied on automated testing.

## Risks

If `{` is not a good delimiter for C functions, this may result in not
properly disambiguating C function bodies. I tested this and at least in
GCC `{` must be used, else it results in a compilation error.

## References

None

## Checklist

- [x] I added tests for this PR's change (or explained in the PR
description why tests don't make sense).